drcachesim trace conversion, extra statistics
drcachesim output still needs testing to see if it actually works.
Extra statistics includes paddr == vaddr checks, as I found that the
reason some paddrs were out of bounds was that `get_physical_address`
in the RISC-V target of QEMU occasionally tells you the paddr is the
same as the vaddr (when there's just been a context switch), which is
completely incorrect. Hopefully they're already being filtered out with
the `check_paddr_valid` checks, but the statistics may show otherwise.