drcachesim conversion: supports vaddr with phys mapping and paddr-only
The "vaddr with phys mapping" trace works with and without the
use_physical option in drcachesim. The "paddr-only" trace is used
in drcachesim without the use_physical option, but produces results
that match the "vaddr with phys mapping" trace with use_physical.

Also renamed program to traceconv, and changed submodule branch (to
reflect the fact that drcachesim-cheri-compatibility was merged into
drcachesim-cheri).