fixed coherency issue for 2 sharers
The fact that some tag bits are unknown presents a bit of an issue.

When we request to read a cache line from cache A, we update the known
bits to reflect this read. The known bits are not typically flushed
until a write back. When we request to read the same cache line from a
peer cache, cache B, we write those known bits back to the closest
common parent of A and B. Since it is the first access from B, the
request will naturally bring in the updated known bits from A into B.
The read access from B will also further update the known bits.

If at this point, the cache line in A is evicted and the known bits
are (invisibly) written back into the closest common parent, the cache
line will not be shared anymore. B will have updated known bits, to be
invisibly written back at a later time. All is good.

However, if the copy in B is written back before the copy in A, when
A tries to write it back, it could fail. This is because the copy in B
may have more known bits set. When A writes it back, it finds that what
it is trying to overwrite is more up to date.

Worse still, if instead of any evictions, A is accessed again, both the
cache line in A and the cache line in B will have (differently) updated
known bits. The copy in A will not have the updates from B. This will
break during one of the write backs (again because we assume the known
bits monotonically increase).

I believe the previous approach worked for our test cases because the
cache line in A is never accessed again, and because the cache line was
written back from A and then B (in that order).

We now perform invisible "read backs" to ensure that the known bits in
A and B match before every access. Not sure this solution will work for
more than 2 sharers, however.