coherency: tried a tag merging approach to simplify things
Not sure that this solves all cases as well. It would be nice if the
unknown tags in the dram request traces correspond to memory that has
not been accessed. If we are not careful to make sure that all the
unknown tags are consolidated before a LLC miss or write back, this
will not hold.