cache coherence: possibly solved issue with multiple sharers
The key to this solution is ensuring that, before every access, all
sharers of a cache line should contain the same state.

We check if any peer caches contain the line we are trying to access
and write back if modified and invalidate if the current access is a
write, as we did previously.

Since reads effectively modify tag state as well (because they uncover
new tags the first time a word is accessed), we have to write those
back as well from the peer caches, so they are read in correctly if
the current access is the first access from this cache.

Most importantly, after a read is performed, if we know that this
cache line was shared, we propagate the updated tag state up until the
closest common parent cache, and down into all of the children where
the cache line exists.